Add Initially Hidden Carousel Bug to KNOWN_ISSUES.MD
Following up on the request to add this to the documentation (noticing that there's a large portion of users who struggle with it and after experiencing myself this seemingly random issue on the latest stable release on a new project), I'm adding all the info I think is relevant here. 

I paraphrased a few of the most common answers, especially those answered by @bd-arc.

Feel free to suggest any modifications to the text or the solutions.